Clean up split test namespaces and helpers
