+++ src/polaris-api-csharpTests/Methods/RequestShape/HoldRequestCancelTests.cs
@@ -1,16 +1,8 @@
 using Clc.Polaris.Api;
 using Clc.Polaris.Api.Tests;
-using Clc.Polaris.Api.Models;
-using Clc.Rest.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Reflection;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace Clc.Polaris.Api.Tests.Methods.RequestShape
